release: Max updates rerelease 0 2 4 (#1423)
* update compiler release

* Fully complete ++ and --

* delete unused files for replacing/converting

* Adding many more tests/examples, update all examples and tests, complete LLVM improvnment and optimize generator

* delete runner.py script inside benchmark, we have salam version so no need py

* updating benchmark salam files

* Adding more benchmark scripts - ideas

* update fa and en books

* Update examples and keep improving compiler to generate better and faster binary and object code, FASTER THAN OTHER LANGS

* Adding newline after import and before main function, update salam files

* Fix bug the Salam slice type (int[:] etc.) compiles down to a 16-byte C struct - Normalize import statements - optimize strlen for all strings, improve performance of working with strings

* updating lexer and semantic to check keywords and identifiers

* Update: langpack_lookup_keyword

* Change and normalize ی/ک arabic into persian

* develop more games in fa

* Adding tests in js mode for en/fa/ar langs

* Update salam.cfg, update std of layout, update std/dom and update compiler semantic

* update enf of dockerfile

* Adding ninja game:

* ignore files that starting with _ in the root

* update jewel and ninja game

* update semantic and fix build error

* update salam files, add lang into ar example files and test files, update dockerfile, bash

* Update

* rename merhaba to hello, update codespell file

* Extend webview and codegen and js driver compiler

* remove old and temp mod files

* update some english names using _ for now

* update webview examples

* layout_gen_element

---------

+++ compiler/src/interp/interp.c
@@ -37,13 +37,13 @@ void tick(interp_t *I)
 
 static size_t itab_hash(const char *s)
 {
-    size_t h = 1469598103934665603u;
+    uint64_t h = 1469598103934665603ULL;
     const unsigned char *p = (const unsigned char *)s;
     for (; *p; p++) {
         h ^= *p;
-        h *= 1099511628211u;
+        h *= 1099511628211ULL;
     }
-    return h;
+    return (size_t)h;
 }
 
 static size_t ptab_hash(const void *p)
@@ -522,14 +522,6 @@ void build_modules(interp_t *I, ast_node_t *program)
     }
 }
 
-/*
- * Generic struct/impl instantiations (e.g. HashMap<str,str>) are deep-cloned
- * by sema (g_instantiate_struct) into whichever file's AST happened to be
- * under analysis at instantiation time, which is not necessarily the file
- * that defines the generic template (e.g. std/collections). Their methods
- * still reference that template's own package-level imports, so they must
- * run with the TEMPLATE's env, not the instantiating file's env.
- */
 static void fixup_generic_envs_in(interp_t *I, vec_t *defs)
 {
     size_t i = 0;
